update apt and dnf package management
