+++ pkg/package_manager/pkgconfig/apt/debian_11.yaml
