+++ pkg/package_manager/pkgconfig/apt/debian_12.yaml
