+++ pkg/package_manager/pkgconfig/apt/debian_6.yaml
