+++ pkg/package_manager/pkgconfig/apt/debian_7.yaml
