+++ pkg/package_manager/pkgconfig/apt/debian_duke.yaml
