+++ pkg/package_manager/pkgconfig/apt/ubuntu_12.04.yaml
