+++ pkg/package_manager/pkgconfig/apt/ubuntu_14.04.yaml
