+++ pkg/package_manager/pkgconfig/apt/ubuntu_22.10.yaml
