+++ pkg/package_manager/pkgconfig/dnf/centos_10.yaml
