+++ pkg/package_manager/pkgconfig/dnf/centos_7.yaml
