+++ pkg/package_manager/pkgconfig/dnf/centos_8.yaml
