+++ pkg/package_manager/pkgconfig/dnf/rocky.yaml
